Rework edge ngram
