+++ Cargo.toml
@@ -27,7 +27,7 @@ embedded_bm = ["rphonetic/embedded_bm"]
 embedded_dm = ["rphonetic/embedded_dm"]
 
 [dependencies]
-tantivy = "0.19"
+tantivy = "0.20"
 
 # Switch to full icu lib when possible
 rust_icu_sys = { version = "4.0", optional = true }
@@ -41,8 +41,11 @@ rphonetic = { version = "2.1", optional = true }
 # Automaton
 fst = { version = "0.4", optional = true }
 
-derive_builder = { version = "0.12", optional = true }
+# Error handling
+thiserror = "1.0"
 
+rustc-hash = "1.1"
+derive_builder = { version = "0.12", optional = true }
 either = {version = "1.7", optional = true}
 
 [patch.crates-io]
